Add Field providing Deck's name and access, remove name from Deck
Created a new class Field which contains name and access permissions for a Deck
instance.
Since a name of a Deck instance is supposed to be used within Field context
only, the name variable was removed from the Deck class body (and, therefore,
Deck constructor was updated and getName() method was removed as well).